1. The last digit

+++ 0x01-variables_if_else_while/1-last_digit.c
@@ -6,26 +6,26 @@
 /* betty style doc for function main goes there */
 int main(void)
 {
-	int n,last;
+	long n, last;
 
 	srand(time(0));
 	n = rand() - RAND_MAX / 2 ;
 	last = n % 10;
-	if(last>5)
+	if (last > 5)
 	{
-		printf("Last digit of %d is %d and is greater than 5",n,last);
+		printf("Last digit of %d is %d and is greater than 5", n, last);
 	}
 	else
 	{
-		if(last==0)
+		if (last == 0)
 		{
-			printf("Last digit of %d is %d and is 0",n,last);
+			printf("Last digit of %d is %d and is 0", n, last);
 		}
 		else
 		{
-			if(last<6)
+			if (last < 6)
 			{
-				printf("Last digit of %d is %d and is less than 6 and not 0",n,last);
+				printf("Last digit of %d is %d and is less than 6 and not 0", n, last);
 			}
 		}
 	}
